+++ src/main/file/FileChunk.java
@@ -2,7 +2,7 @@ package main.file;
 
 public interface FileChunk {
     int size();
-    int get(int i);
+    byte get(int i);
 
     byte[] asByteArray();
 }
